refactor: like ttl 상수화
